Consolidate common code in PauliObject interface
